refactor: move unroll_sequences in unrolling.py
