+++ src/qibolab/platform/__init__.py
@@ -1,4 +1,4 @@
 from .load import create_platform
-from .platform import Platform, unroll_sequences
+from .platform import Platform
 
-__all__ = ["Platform", "create_platform", "unroll_sequences"]
+__all__ = ["Platform", "create_platform"]
