+++ src/qibolab/platform/platform.py
@@ -15,7 +15,6 @@ from qibolab.execution_parameters import ExecutionParameters
 from qibolab.identifier import ChannelId
 from qibolab.instruments.abstract import Controller, Instrument, InstrumentId
 from qibolab.parameters import NativeGates, Parameters, Settings, update_configs
-from qibolab.pulses import Delay
 from qibolab.qubits import Qubit, QubitId, QubitPairId
 from qibolab.sequence import PulseSequence
 from qibolab.sweeper import ParallelSweepers
@@ -36,36 +35,6 @@ T = TypeVar("T")
 def default(value: Optional[T], default: T) -> T:
     """None replacement shortcut."""
     return value if value is not None else default
-
-
-def unroll_sequences(
-    sequences: list[PulseSequence], relaxation_time: int
-) -> tuple[PulseSequence, dict[int, list[int]]]:
-    """Unrolls a list of pulse sequences to a single sequence.
-
-    The resulting sequence may contain multiple measurements.
-
-    `relaxation_time` is the time in ns to wait for the qubit to relax between playing
-    different sequences.
-
-    It returns both the unrolled pulse sequence, and the map from original readout pulse
-    serials to the unrolled readout pulse serials. Required to construct the results
-    dictionary that is returned after execution.
-    """
-    total_sequence = PulseSequence()
-    readout_map = defaultdict(list)
-    for sequence in sequences:
-        total_sequence.concatenate(sequence)
-        # TODO: Fix unrolling results
-        for _, acq in sequence.acquisitions:
-            readout_map[acq.id].append(acq.id)
-
-        length = sequence.duration + relaxation_time
-        for channel in sequence.channels:
-            delay = length - sequence.channel_duration(channel)
-            total_sequence.append((channel, Delay(duration=delay)))
-
-    return total_sequence, readout_map
 
 
 def estimate_duration(
